chatbot_/requirements.txt
altgraph==0.17
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
coverage==5.0.4
hvac==0.10.0
idna==2.9
macholib==1.14
modulegraph==0.18
Pillow==7.1.2
pony==0.7.13
psycopg2-binary==2.8.5
pycparser==2.20
requests==2.23.0
six==1.14.0
urllib3==1.25.8
vk-api==11.7.0
